dep-flask-front/pyproject.toml
[project]
name = "dep-flask-front"
version = "0.1.0"
description = "Flask fronted for depot project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[tool.pyright]
venvPat = "."
venv = ".venv"
